// application/typescript/library/amplify/backend/function/eventLogStreamProcessor/lib/library/codegen/API.ts
export enum PlayerStatus {
  withdrew = "withdrew",
  joined = "joined",
  winner = "winner",
}
// ...
export type PoolPlayer = {
  __typename: "PoolPlayer",
  id: string,
  userWallet: UserWallet,
  status: PlayerStatus,
  pool: Pool,
  createdAt: string,
  updatedAt: string,
  poolPlayersId?: string | null,
  poolPlayerUserWalletId: string,
};

export type UserWallet = {
  __typename: "UserWallet",
  id: string,
  wallet: string,
  nickname?: string | null,
  chatlogo?: string | null,
  brands?: ModelBrandConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelBrandConnection = {
  __typename: "ModelBrandConnection",
  items:  Array<Brand | null >,
  nextToken?: string | null,
};

export type Brand = {
  __typename: "Brand",
  id: string,
  wallet: string,
  name: string,
  rating: number,
  logo?: string | null,
  completedPools: number,
  failedPools: number,
  createdAt: string,
  updatedAt: string,
  userWalletBrandsId?: string | null,
};

export type Pool = {
  __typename: "Pool",
  id: string,
  poolId: number,
  poolTitle: string,
  poolCategory: PoolCategory,
  poolCreator: UserWallet,
  poolType: poolType,
  poolStatus: string,
  poolEntryFee: string,
  poolTotal: string,
  poolWinningPayout: string,
  allowPlayerLeave: boolean,
  apiPoolAttributes?: ApiPoolAttributes | null,
  lotteryPoolAttributes?: LotteryPoolAttributes | null,
  players?: ModelPoolPlayerConnection | null,
  createdAt: string,
  updatedAt: string,
  poolPoolCreatorId: string,
  poolApiPoolAttributesId?: string | null,
  poolLotteryPoolAttributesId?: string | null,
};

export enum PoolCategory {
  other = "other",
  sports = "sports",
  esports = "esports",
  gaming = "gaming",
  politics = "politics",
  financial = "financial",
  random_winners = "random_winners",
}


export enum poolType {
  lottery = "lottery",
  manual = "manual",
  api = "api",
  custom_builder = "custom_builder",
}


export type ApiPoolAttributes = {
  __typename: "ApiPoolAttributes",
  id: string,
  apiKey: string,
  lockFundsDatetime: string,
  executeWinnerDatetime: string,
  apiUrlForResults: string,
  verifiedUrlSchema: boolean,
  poolCreatorPercentFeeToWithold: number,
  apiWinnerOptionEntries: string,
  apiwinnerResult: string,
  poolApiDefaultSchema: string,
  poolApiCustomSchema?: string | null,
  pool?: Pool | null,
  createdAt: string,
  updatedAt: string,
  apiPoolAttributesPoolId?: string | null,
};

export type LotteryPoolAttributes = {
  __typename: "LotteryPoolAttributes",
  id: string,
  auditRecordDrawId?: string | null,
  isAuditEnabled: boolean,
  randomOrgUrlForResults?: string | null,
  pool: Pool,
  createdAt: string,
  updatedAt: string,
  lotteryPoolAttributesPoolId: string,
};

export type ModelPoolPlayerConnection = {
  __typename: "ModelPoolPlayerConnection",
  items:  Array<PoolPlayer | null >,
  nextToken?: string | null,
};
// ...
export type UpdateLotteryPoolAttributesInput = {
  id: string,
  auditRecordDrawId?: string | null,
  isAuditEnabled?: boolean | null,
  randomOrgUrlForResults?: string | null,
  lotteryPoolAttributesPoolId?: string | null,
};
// ...
export type CreateUserWalletInput = {
  id?: string | null,
  wallet: string,
  nickname?: string | null,
  chatlogo?: string | null,
};